pyproject.toml
[project]
name = "powerpit"
version = "0.1.0"
description = "Power Pit video generator (M0 setup)"
authors = [{ name = "Power Pit" }]
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pillow",
    "imageio[ffmpeg]",
    "pyyaml",
    "pymunk",
    "moviepy",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"
